Add Type XRefs to IR (#51)
This updates the IR representation of types to a possible list of strings or
xref objects. The xrefs can be internal xrefs to types in the project or
external ones to types defined in dependencies. I don't currently add xrefs
to intrinsic types but that would be a natural thing to do.

The renderer currently throws this information away. Rendering it is left to a
followup.