Improve typedoc path handling a bit (#72)
We can find a path relative to the base of the github repo by looking at
`sources.url`. This isn't ideal, it would be nice if typedoc just included an
absolute path somewhere. On the other hand, it's better than nothing.

This also cleans up some dead code that supported old typedoc versions